nuevos cambios
voy de a poco pero voy andando

+++ src/main/java/com/example/Homebanking/Entidades/Transferencia.java
@@ -39,6 +39,6 @@ public class Transferencia {
     Date Fecha;
 
     @Column(nullable = false)
-    Integer monto;
+    Double monto;
 
 }
